(with-server) On purchase, wait until webhook have been processed by server
When using iaptic, the status of the purchase will only be correct after the
server had received the webhook call containing purchase information.

Thus, the client should wait for this webhook to have been received.

After a purchase, call the "pending-webhook" endpoint on the server so it knows
that a webhook should happen for that user that will update the subscription
status.

When there's a pending webhook, the UI is reflecting the fact that purchase is
still being processed.